fix: make `isFalsyType()` return `true` for `0n` (#545)
<!-- 👋 Hi, thanks for sending a PR to ts-api-utils! 💖.
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #544 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/ts-api-utils/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/ts-api-utils/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview
The `type.value` returns a `PseudoBigInt` object rather than an actual
bigint for TS legacy support reasons. Therefore, to determine if a
literal is a falsy value, we cannot rely on the truthiness of
`type.value` alone, and need to check what the actual object is if it's
not a primitive.