docs: deprecate symbolHasReadonlyDeclaration (#766)
## PR Checklist

- [x] Addresses an existing open issue: fixes #610
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/ts-api-utils/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/ts-api-utils/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Adds a `@deprecated` JSDocs tag. This is purely a docs change for now -
it does not remove the function. The function is actually used by
`isReadonlyPropertyIntersection`, so the next step would be to remove
the `export` - not the whole function.

💖